07-ia/ia-ii/requirements.txt
fastmcp==3.1.1
python-dotenv>=1.0.0
rich>=13.0.0
agent-framework-core==1.0.0rc5